refactor: use tailwind-variants for styling
