+++ design-toolkit/components/package.json
@@ -64,11 +64,12 @@
     "@fontsource/roboto-mono": "^5.1.0",
     "@internationalized/date": "^3.8.0",
     "@radix-ui/react-avatar": "^1.1.4",
+    "@react-aria/ssr": "^3.9.6",
     "@react-aria/utils": "^3.29.1",
     "@react-stately/color": "^3.8.6",
+    "@react-stately/data": "^3.13.1",
     "@react-stately/datepicker": "^3.14.2",
     "@react-stately/utils": "^3.10.7",
-    "@react-aria/ssr": "^3.9.6",
     "@react-types/shared": "^3.25.0",
     "@tanstack/react-table": "^8.21.3",
     "react": "^19.0.0",
@@ -81,7 +82,6 @@
   "dependencies": {
     "@accelint/core": "workspace:*",
     "@accelint/icons": "workspace:*",
-    "@react-stately/data": "3.13.1",
     "client-only": "0.0.1",
     "clsx": "^2.1.1",
     "cva": "npm:class-variance-authority@^0.7.1",
@@ -97,6 +97,7 @@
     "@accelint/typescript-config": "workspace:*",
     "@accelint/vitest-config": "workspace:*",
     "@react-stately/color": "3.8.6",
+    "@react-stately/data": "3.13.1",
     "@react-stately/datepicker": "3.13.0",
     "@react-stately/utils": "3.10.7",
     "@react-types/shared": "3.25.0",
