+++ design-toolkit/components/src/components/table/types.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import type {
+  ForwardedRef,
+  HTMLAttributes,
+  RefAttributes,
+  TableHTMLAttributes,
+  TdHTMLAttributes,
+  ThHTMLAttributes,
+} from 'react';
+import type { VariantProps } from 'tailwind-variants';
+import type { cellStyles, headerCellStyles } from './styles';
+
+export type TableProps = TableHTMLAttributes<HTMLTableElement> &
+  RefAttributes<HTMLTableElement>;
+
+export type TableBodyProps = HTMLAttributes<HTMLTableSectionElement> &
+  RefAttributes<HTMLTableSectionElement>;
+
+export type TableRowProps = HTMLAttributes<HTMLTableRowElement> &
+  RefAttributes<HTMLTableRowElement>;
+
+export type TableCellProps = TdHTMLAttributes<HTMLTableCellElement> &
+  VariantProps<typeof cellStyles> & {
+    ref?: ForwardedRef<HTMLTableCellElement>;
+  };
+
+export type TableHeaderCellProps = ThHTMLAttributes<HTMLTableCellElement> &
+  VariantProps<typeof headerCellStyles> &
+  RefAttributes<HTMLTableCellElement>;
+
+export type TableHeaderProps = HTMLAttributes<HTMLTableSectionElement> &
+  RefAttributes<HTMLTableSectionElement>;
